Add files via upload
Codes for part 1 and 2. The result of the analysis can be gotten when the right generation data . one of the codes shows data for solar PV farm analysis(1330) and the 2140 code shows the distribution network of the Balearic Islands at 21:40pm. Canada code is for the second part of the assignment.